+++ android/src/main/java/com/freeraspreactnative/events/BaseRaspEvent.kt
@@ -4,7 +4,4 @@ internal interface BaseRaspEvent {
   val value: Int
   val channelName: String
   val channelKey: String
-
 }
-
-